Punto de Conexion (y otras cositas)

+++ model/punto_conexion.php
@@ -10,8 +10,12 @@ class punto_conexion extends Conexion {
     private $puerto_patch_panel;
 
     public function __construct() {
-        $this->conex = new Conexion("sistema");
-        $this->conex = $this->conex->Conex();
+
+        $this->id_punto_conexion = "";
+        $this->codigo_patch_panel = "";
+        $this->id_equipo = "";
+        $this->puerto_patch_panel = "";
+
     }
 
 
@@ -49,12 +53,17 @@ class punto_conexion extends Conexion {
 
         try {
 
+            $this->conex = new Conexion("sistema");
+            $this->conex = $this->conex->Conex();
+            $this->conex->beginTransaction();
+
             $query = "SELECT * FROM punto_conexion WHERE codigo_patch_panel = :codigo_patch_panel AND puerto_patch_panel = :puerto_patch_panel";
 
             $stm = $this->conex->prepare($query);
             $stm->bindParam(":codigo_patch_panel", $this->codigo_patch_panel);
             $stm->bindParam(":puerto_patch_panel", $this->puerto_patch_panel);
             $stm->execute();
+            $this->conex->commit();
 
             if ($stm->rowCount() > 0) {
                 $dato['arreglo'] = $stm->fetch(PDO::FETCH_ASSOC);
@@ -65,17 +74,22 @@ class punto_conexion extends Conexion {
 
         } catch (PDOException $e) {
 
+            $this->conex->rollBack();
+            $dato['bool'] = -1;
             $dato['error'] = $e->getMessage();
 
         }
 
-        $this->Cerrar_Conexion($none, $stm);
+        $this->Cerrar_Conexion($this->conex, $stm);
 
         return $dato;
-    } //validar punto conexion, puerto -- patch panel
+    } 
         
 
     private function Registrar() {
+
+        $this->conex = new Conexion("sistema");
+        $this->conex = $this->conex->Conex();
 
         $dato = [];
         $bool = $this->Validar();;
@@ -83,6 +97,8 @@ class punto_conexion extends Conexion {
         if ($bool['bool'] == 0) {
 
             try {
+
+                $this->conex->beginTransaction();
 
                 $query = "INSERT INTO punto_conexion(codigo_patch_panel, id_equipo,  puerto_patch_panel) VALUES 
                 (:codigo_patch_panel, :id_equipo, :puerto_patch_panel)";
@@ -93,12 +109,16 @@ class punto_conexion extends Conexion {
                 $stm->bindParam(":id_equipo", $this->id_equipo);
                 $stm->bindParam(":puerto_patch_panel", $this->puerto_patch_panel);
                 $stm->execute();
+
+                $this->conex->commit();
+
                 $dato['resultado'] = "registrar";
                 $dato['estado'] = 1;
                 $dato['mensaje'] = "Se Registro el Punto de Conexión Exitosamente";
 
             } catch (PDOException $e) {
 
+                $this->conex->rollBack();
                 $dato['resultado'] = "error";
                 $dato['estado'] = -1;
                 $dato['mensaje'] = $e->getMessage();
@@ -107,6 +127,7 @@ class punto_conexion extends Conexion {
 
         } else {
 
+            $this->conex->rollBack();
             $dato['resultado'] = "error";
             $dato['estado'] = -1;
             $dato['mensaje'] = "Punto de Conexión Ocupado";
@@ -120,17 +141,23 @@ class punto_conexion extends Conexion {
 
     private function Actualizar() {
 
+        $this->conex = new Conexion("sistema");
+        $this->conex = $this->conex->Conex();
+
         $dato = [];
-        
+        $bool = $this->Validar();
 
-        
+        if ($bool['bool'] != 0) {
 
-        try {
+            try {
+
+                $this->conex->beginTransaction();
+
                 $query = "UPDATE punto_conexion 
-                            SET codigo_patch_panel = :codigo_patch_panel, 
-                            id_equipo = :id_equipo, 
-                            puerto_patch_panel = :puerto_patch_panel
-                        WHERE id_punto_conexion = :id_punto_conexion";
+                SET codigo_patch_panel = :codigo_patch_panel, 
+                id_equipo = :id_equipo, 
+                puerto_patch_panel = :puerto_patch_panel
+                WHERE id_punto_conexion = :id_punto_conexion";
 
                 $stm = $this->conex->prepare($query);
                 $stm->bindParam(":codigo_patch_panel", $this->codigo_patch_panel);
@@ -139,24 +166,29 @@ class punto_conexion extends Conexion {
                 $stm->bindParam(":id_punto_conexion", $this->id_punto_conexion);
                 $stm->execute();
 
+                $this->conex->commit();
+
                 $dato['resultado'] = "modificar";
                 $dato['estado'] = 1;
                 $dato['mensaje'] = "Se modificó el Punto de Conexión exitosamente";
-                $this->Cerrar_Conexion($this->conex, $stm);
-
+            
             } catch (PDOException $e) {
 
+                $this->conex->rollBack();
                 $dato['estado'] = -1;
                 $dato['resultado'] = "error";
                 $dato['mensaje'] = $e->getMessage();
 
             }
-        
+        } else {
 
+            $this->conex->rollBack();
+            $dato['resultado'] = "error";
+            $dato['estado'] = -1;
+            $dato['mensaje'] = "Error al Modificar el registro";
+
+        }
             
-
-        
-
         $this->Cerrar_Conexion($this->conex, $stm);
 
         return $dato;
@@ -164,33 +196,45 @@ class punto_conexion extends Conexion {
 
     private function Eliminar() {
 
+        $this->conex = new Conexion("sistema");
+        $this->conex = $this->conex->Conex();
+
         $dato = [];
-        
+        $bool = $this->Validar();
 
+        if ($bool['bool'] != 0) {
         
-
             try {
+
+                $this->conex->beginTransaction();
+
                     $query = "DELETE FROM punto_conexion WHERE id_punto_conexion = :id_punto_conexion";
 
                 $stm = $this->conex->prepare($query);
                 $stm->bindParam(":id_punto_conexion", $this->id_punto_conexion);
                 $stm->execute();
 
+                $this->conex->commit();
+
                 $dato['resultado'] = "eliminar";
                 $dato['estado'] = 1;
                 $dato['mensaje'] = "Se eliminó el Punto de Conexión exitosamente";
-                $this->Cerrar_Conexion($this->conex, $stm);
 
             } catch (PDOException $e) {
 
+                $this->conex->rollBack();
                 $dato['resultado'] = "error";
                 $dato['estado'] = -1;
                 $dato['mensaje'] = $e->getMessage();
 
             }
+        } else {
 
-       
+            $dato['resultado'] = "error";
+            $dato['estado'] = -1;
+            $dato['mensaje'] = "Error al eliminar el registro";
 
+        }
 
         $this->Cerrar_Conexion($this->conex, $stm);
 
@@ -199,16 +243,19 @@ class punto_conexion extends Conexion {
 
     private function Consultar() {
 
+        $this->conex = new Conexion("sistema");
+        $this->conex = $this->conex->Conex();
+
         $dato = [];
 
         try {
 
             $query = "SELECT 
-                        id_punto_conexion, 
-                        codigo_patch_panel, 
-                        id_equipo, 
-                        puerto_patch_panel
-                    FROM punto_conexion";
+            id_punto_conexion, 
+            codigo_patch_panel, 
+            id_equipo, 
+            puerto_patch_panel
+            FROM punto_conexion";
 
             $stm = $this->conex->prepare($query);
             $stm->execute();
